### 2024-05-27 #6778 @loicdtx
Dear @kthyng thanks for handling this pre-review.

## About potential reviewers

From the list of people that have agreed to review I have identified two potential good fit:
- Tobias Seydewitz (tobijjah)
- Sarah Zwiep (szwiep)

And another not in the list:
- Alex Leith (alexgleith)

## Regarding the error

I double checked everything (repository is public and url is correct, branch name is correct, manuscript builds flawlessly using the docker image) and couldn't find the source of the issue. Could it just be bad luck?
