### 2024-05-27 #6778 @loicdtx
Build is still failing at the download/clone stage. I explored https://github.com/xuanxu/theoj/blob/main/lib/theoj/paper.rb and https://github.com/xuanxu/paper-action/blob/main/get_paper.rb but couldn't figure out anything obvious; I also don't know much ruby.
I tried cloning fully anonymously (from a clean docker container) and that worked fine (`git clone -b joss-paper https://github.com/ec-jrc/nrt`)
