### 2024-05-28 #6778 @kthyng
@openjournals/dev Could you help debug this paper compilation error? Thank you!
