### 2024-05-30 39d2151
Add DOIs to most bib entries. Remove uncited references
