+++ paper/refs.bib
@@ -4,7 +4,8 @@
   journal={Econometrica},
   pages={47--78},
   year={1998},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.2307/2998540}
 }
 
 @article{brooks2013fly,
@@ -15,7 +16,8 @@
   number={6},
   pages={3316--3332},
   year={2013},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TGRS.2013.2272545}
 }
 
 @inproceedings{buitinck2013api,
@@ -29,6 +31,7 @@
   booktitle = {ECML PKDD Workshop: Languages for Data Mining and Machine Learning},
   year      = {2013},
   pages = {108--122},
+  doi = {10.48550/arXiv.1309.0238}
 }
 
 @article{bullock2020improved,
@@ -38,7 +41,8 @@
   volume={238},
   pages={111165},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.rse.2019.04.018}
 }
 
 @article{bullock2022timeliness,
@@ -48,7 +52,8 @@
   volume={276},
   pages={113043},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.rse.2022.113043}
 }
 
 @book{foresteurope2020state,
@@ -68,32 +73,23 @@
   year={2019}
 }
 
-@article{kennedy2010detecting,
-  title={{Detecting trends in forest disturbance and recovery using yearly Landsat time series: 1. LandTrendr—Temporal segmentation algorithms}},
-  author={Kennedy, Robert E and Yang, Zhiqiang and Cohen, Warren B},
-  journal={Remote Sensing of Environment},
-  volume={114},
-  number={12},
-  pages={2897--2910},
-  year={2010},
-  publisher={Elsevier}
-}
-
 @inproceedings{killough2018overview,
   title={Overview of the open data cube initiative},
   author={Killough, Brian},
   booktitle={IGARSS 2018-2018 IEEE international geoscience and remote sensing symposium},
   pages={8629--8632},
   year={2018},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/IGARSS.2018.8517694}
 }
 
 @inproceedings{lam2015numba,
-  title={{Numba: A llvm-based python jit compiler}},
+  title={{Numba: A LLVM-based python JIT compiler}},
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
-  year={2015}
+  year={2015},
+  doi={10.1145/2833157.2833162}
 }
 
 @article{pedregosa2011scikit,
@@ -103,18 +99,8 @@
   volume={12},
   pages={2825--2830},
   year={2011},
-  publisher={JMLR. org}
-}
-
-@article{puhm2020near,
-  title={{A near real-time method for forest change detection based on a structural time series model and the Kalman filter}},
-  author={Puhm, Martin and Deutscher, Janik and Hirschmugl, Manuela and Wimmer, Andreas and Schmitt, Ursula and Schardt, Mathias},
-  journal={Remote Sensing},
-  volume={12},
-  number={19},
-  pages={3135},
-  year={2020},
-  publisher={MDPI}
+  publisher={JMLR. org},
+  doi={10.48550/arXiv.1201.0490}
 }
 
 @article{soille2018versatile,
@@ -147,7 +133,8 @@
   volume={123},
   pages={98--108},
[…]
