### 2024-05-30 #6778 @loicdtx
Thanks everyone, editors and reviewers; great to see that review moving so fast!
In the meantime I have added missing DOIs to most bib entries as suggested by the review bot.
