### 2024-05-30 #6778 @alexgleith
What is the practical implementation of the review checklist? Do I create a comment and check through the items? Or does that get added to the issue up the top and I complete it there?
