### 2024-05-31 #6815 @editorialbot
[REVIEW]: NRT: operational monitoring of satellite image time-series in python
**Submitting author:** @loicdtx (<a href="http://orcid.org/0000-0002-5058-2526">Loïc Dutrieux</a>)
**Repository:** https://github.com/ec-jrc/nrt
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.2.1
