### 2024-06-03 853a288
Minor change, suggestion from JRC internal review (consistently write nrt in lowercase)
