### 2024-06-03 #6815 @loicdtx
Thanks @AdamRJensen for kicking off the review.

- Recompiling pdf after very minor change to the manuscript https://github.com/ec-jrc/nrt/commit/853a288a93b24bc101c00f2644080e0b059e2f04
- Tagging @jonasViehweger to keep him updated on the review progress
