### 2024-06-24 #6815 @alexgleith
There are a couple of issue that I think it would be good to resolve, @loicdtx. One is easy, by adding something to the readme about reporting issues and contributing, or adding a 'contributing.md' file.

The other is around reproducibility of the examples. I'd actually like an example that streamed data too, rather than using the dumped NetCDF files, but this might be out of scope and too much effort. Pinning xarray or other library versions so that the sample code works for others (like me!) will be enough to get this over the line for me.
