### 2024-06-24 d82df88
Bug in quickstart example. Dimensions names must be explicitely passed to xarray.DataArray.sel fixes #11
