### 2024-06-26 a05414f
New example in sphinx doc gallery. Load data from STAC, create forest mask and run IQR
