+++ docs/gallery/dontplot_bolivia_stac.py
@@ -1,0 +1,324 @@
+r"""
+STAC monitoring pipeline
+========================
+
+*Run a monitoring simulation anywhere in the world using STAC*
+
+This example demonstrates how to retrieve satellite image time-series
+from a cloud archive, given a pair of coordinates anywhere in the world,
+and use them for a monitoring simulation. Here, “simulation” does not
+refer to simulated data; the example uses real data to detect actual
+land dynamics. The term highlights the retrospective nature of the
+exercise, distinguishing it from real-time monitoring systems. For this
+example, data is retrieved from the `Microsoft Planetary Computer
+archive <https://planetarycomputer.microsoft.com/catalog>`__, which is
+indexed in a Spatio-Temporal Asset Catalogue (STAC), making it queryable
+in an interoperable way. While querying and preparing data extends
+slightly beyond the ``nrt`` package’s scope, we aim to provide a
+comprehensive view of the entire pipeline. Specifically, this example
+demonstrates how to:
+
+-  Query Landsat Collection 2 data for any location using the Planetary
+   Computer STAC catalogue
+-  Load a datacube as an in-memory xarray dataset using ``odc-stac`` and
+   ``dask``
+-  Prepare the Landsat data for further analysis (mask clouds and
+   shadows, apply scaling and offset, compute vegetation indices)
+-  Generate a simple rule-based forest mask
+-  Simulate a near real-time monitoring scenario
+"""
+# sphinx_gallery_thumbnail_path = 'images/iqr_results_bolivia.png'
+
+#############################################
+# Define the study area
+# ---------------------
+# 
+# Here we prepare a bounding box of 20 by 20 km, centered on a forest area
+# near the city of Concepción in Bolivia. Center coordinates are in
+# longitude, latitude which is the Coordinate Reference System we
+# generally get from most online services (google maps or simply typing
+# “name of the place coordinates” in a search engine). To expand these
+# coordinates to a bounding box of 20 km side, we transform the center
+# point to a metric equidistant custom CRS (local UTM, if known could be
+# used as well) and expand the point. Because the STAC API, later used to
+# query the data, expects a bounding box with coordinates in EPSG:4326, we
+# convert the bbox back to that CRS.
+from shapely.geometry import Point
+from shapely.ops import transform
+from pyproj import CRS, Transformer
+
+# Define centroid as a shapely point (in EPSG:4326)
+centroid = Point(-61.725, -16.162)
+# Custom equidistant CRS centered on the centroid
+local_crs = CRS(proj='aeqd', datum='WGS84', lat_0=centroid.y, lon_0=centroid.x)
+transformer = Transformer.from_crs(CRS.from_epsg(4326), local_crs, always_xy=True)
+# Transform centroid to custom CRS, expand to bbox and transform back to EPSG:4326
+centroid_proj = transform(transformer.transform, centroid)
+bbox_proj = centroid_proj.buffer(10000).bounds
+bbox = transformer.transform_bounds(*bbox_proj, direction='INVERSE')
+
+#######################################################
+# Query and load the data
+# -----------------------
+# 
+# Here we query 5 years of Landsat data that intersect the previously
+# defined study area. ``pystac-client`` is used to handle the query, while
+# we use ``odc-stac`` in combination with ``dask`` to efficiently load the
+# data as an ``xarray.Dataset``. The local Dask cluster allows loading
+# data in parallel, which can speed up the overall process. However, the
+# main bottleneck here is network speed, and parallelization may not be
[…]
