+++ docs/conf.py
@@ -43,6 +43,7 @@ extensions = [
 
 # Gallery configuration
 sphinx_gallery_conf = {
+     'filename_pattern': '/plot_',
      'examples_dirs': 'gallery',   # path to your example scripts
      'gallery_dirs': 'auto_examples',  # path to where to save gallery generated output
 }
@@ -70,7 +71,7 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ['_static', 'images']
 html_logo = "images/logo.png"
 html_theme_options = {
     'logo_only': True,
