### 2024-06-26 fe8a868
Decrease subtitles by one level in gallery examples to get a coherent gallery TOC
