### 2024-06-26 e5ee261
Add references to STAC gallery example
