### 2024-06-26 0e7eb48
Minor textual changes to STAC gallery example
