### 2024-06-26 8adf420
Add contributing guidelines
