### 2024-06-26 #6815 @loicdtx
Thank you for your review, @alexgleith, and for the valuable suggestions. I've incorporated them, and the project now includes a CONTRIBUTING.md file. Additionally, the online documentation now features a [new example](https://nrt.readthedocs.io/en/latest/auto_examples/dontplot_bolivia_stac.html#sphx-glr-auto-examples-dontplot-bolivia-stac-py) that fetches data directly from a cloud archive. This was quite a bit of work, but it was definitely worth it. We've had this in the pipeline for a while, and it just needed this little push.
