### 2024-07-08 #6815 @AdamRJensen
Hi @szwiep - at JOSS we aim to have reviews completed within 4-6 weeks. I hope you're able to start soon on the checklist (the first items are easy and a great way to get started).
