### 2024-07-10 ec-jrc/nrt#13 @szwiep
Missing documentation for tests
Unless I've missed it somewhere, there doesn't seem to be documentation on how to run your tests.

It would be helpful to describe, briefly, how users should execute your unit tests. Typically, I've seen test instructions in the README.md (like [odc-stac](https://github.com/opendatacube/odc-stac?tab=readme-ov-file#developing)) or in the CONTRIBUTING.md (like [rasterio](https://github.com/rasterio/rasterio/blob/main/CONTRIBUTING.rst)).
