### 2024-07-10 ec-jrc/nrt#14 @szwiep
Integration tests failing in Pangeo environments
The integration tests are failing on `xr.open_dataset` calls when I'm using a Pangeo docker image in JupyterHub (2024.07.08):
```
>       raise ValueError(error_msg)
E       ValueError: did not find a match in any of xarray's currently installed IO backends ['netcdf4', 'h5netcdf', 'scipy', 'argo', 'cfgrib', 'gini', 'kerchunk', 'pydap', 'rasterio', 'zarr']. Consider explicitly selecting one of the installed engines via the ``engine`` parameter, or installing additional IO dependencies, see:
E       https://docs.xarray.dev/en/stable/getting-started-guide/installing.html
E       https://docs.xarray.dev/en/stable/user-guide/io.html
```
However, I don't run into any errors when running tests locally on a Windows 11 machine. For both systems, I was running `pytest` from the top level directory of the repo.

It's possible I'm not following the correct testing procedure for your repo, so addressing #13 first might resolve this!
