### 2024-07-10 ec-jrc/nrt#15 @szwiep
Missing documentation on dependancies
JOSS guidelines require a clearly-stated list of dependancies. I can peruse the repo and see the dependancies in your `setup.py` but I think a written list either in your README.md (see [rasterio](https://github.com/rasterio/rasterio?tab=readme-ov-file#rasterio)) or in your documentation page (see [Xarray](https://docs.xarray.dev/en/stable/getting-started-guide/installing.html#required-dependencies)) would give users a quicker snapshot of which libraries your package is built on.
