### 2024-07-10 #6815 @szwiep
Thanks for the ping @AdamRJensen and for your patience @loicdtx. I've completed my first pass through the project and added some issues:
- https://github.com/ec-jrc/nrt/issues/13
- https://github.com/ec-jrc/nrt/issues/14
- https://github.com/ec-jrc/nrt/issues/15

This is super exciting software, even more so because your team is putting it out into the open-source world. Looking forward to moving it through the review process!
