### 2024-07-10 ec-jrc/nrt#14 @loicdtx
@szwiep thanks for reporting this. Is this what you tried?

```bash
# Enter container in interactive mode
docker run --rm -it pangeo/pangeo-notebook:latest /bin/bash
# Within container
git clone https://github.com/ec-jrc/nrt.git
cd nrt
pip install -e .[tests]
pytest
```

I confirm that this results in a failure of the integration tests. The reason is that the image does not include the `git lfs` extension which is used to version binary files accessible through the `nrt.data` module and used in these integration tests.

The following sequence of commands works.

```bash
# Enter container as root
docker run --rm -it --user root pangeo/pangeo-notebook:latest /bin/bash
apt update
apt install git-lfs
git lfs install
git clone https://github.com/ec-jrc/nrt.git
cd nrt
pip install -e .[tests]
pytest
```

I'll add a note about git lfs in the CONTRIBUTING.md. Thanks
