### 2024-07-10 fc6448a
Some fixes to address improvements suggested during JOSS review - Warning note on git lfs added to CONTRIBUTING.md (fixes #14) - List of main dependencies added to README (fixes #15) - Document how to run unit/integration tests in CONTRIBUTING (fixes #13)
