### 2024-07-10 206eee3
Add citing instruction in readme + cff file (fixes #16)
