### 2024-07-10 #6815 @loicdtx
Thanks for your review as well as the clear and constructive issues @szwiep . I believe I've addressed them all; let me know if you think something isn't right or missing.
