### 2024-07-10 ec-jrc/nrt#14 @szwiep
The `git lfs` solution works for me. Thanks for getting to the bottom of this!
