### 2024-07-15 9348c58
Bump version; new release to be archived on Zenodo and linked to the JOSS manuscript "nrt: operational monitoring of satellite image time-series in Python"
