### 2024-07-24 99a8492
Update python to major version only in setup.py
