+++ setup.py
@@ -36,9 +36,7 @@ setup(name='nrt',
       url='https://github.com/ec-jrc/nrt.git',
       license='EUPL-v1.2',
       classifiers=[
-          'Programming Language :: Python :: 3.9',
-          'Programming Language :: Python :: 3.10',
-          'Programming Language :: Python :: 3.11',
+          'Programming Language :: Python :: 3',
       ],
       packages=find_packages(),
       package_data={'nrt': ['data/*.nc', 'data/*.tif', 'data/*.json']},
