### 2024-07-24 174a129
Add python 3.12 to test matrix
