### 2024-07-24 #6815 @AdamRJensen
It would be nice if the nrt acronym was mentioned (near real time)

Also, I've opened a number of issues and PRs in the repository that I hope you're able to address.

Otherwise, things are looking good! 🥳
