### 2024-07-24 0adfc87
Add explicit release date to each version in CHANGE.txt fixes #18
