### 2024-07-24 1d27ab1
fix PR on testing matrix (python 3.12 instead 0f 3.12)
