### 2024-07-24 4533224
Reorganize tests in a single `tests` directory at the project root as suggested by @AdamRJensen Fixes #21
