+++ pytest.ini
@@ -8,4 +8,3 @@ markers =
     ccdc: All test relating to CCDC monitoring
 testpaths =
     tests
-    integration_tests
