### 2024-07-25 d22806b
Add some missing DOIs in readme and docs Fixes #22 and #30
