### 2024-07-25 95204d6
untest fix to bat file for compiling doc on windows fixes #28
