### 2024-07-29 018db32
Re-organized API doc and regroup it into a single file - Simplified tree - Renamed API Reference instead of simply nrt - Fixes #17 - Fixes #19
