+++ docs/conf.py
@@ -71,7 +71,7 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static', 'images']
+html_static_path = ['images']
 html_logo = "images/logo.png"
 html_theme_options = {
     'logo_only': True,
