+++ docs/nrt.data.rst
@@ -1,10 +1,0 @@
-nrt.data package
-================
-
-Module contents
----------------
-
-.. automodule:: nrt.data
-   :members:
-   :undoc-members:
-   :show-inheritance:
