+++ docs/nrt.rst
@@ -1,70 +1,0 @@
-nrt package
-===========
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   nrt.data
-   nrt.monitor
-
-Submodules
-----------
-
-nrt.fit\_methods module
------------------------
-
-.. automodule:: nrt.fit_methods
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nrt.log module
---------------
-
-.. automodule:: nrt.log
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nrt.outliers module
--------------------
-
-.. automodule:: nrt.outliers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nrt.stats module
-----------------
-
-.. automodule:: nrt.stats
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nrt.utils module
-----------------
-
-.. automodule:: nrt.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nrt.utils\_efp module
----------------------
-
-.. automodule:: nrt.utils_efp
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: nrt
-   :members:
-   :undoc-members:
-   :show-inheritance:
