+++ docs/nrt.monitor.rst
@@ -1,53 +1,0 @@
-nrt.monitor package
-===================
-
-Submodules
-----------
-
-nrt.monitor.ccdc module
------------------------
-
-.. automodule:: nrt.monitor.ccdc
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nrt.monitor.cusum module
-------------------------
-
-.. automodule:: nrt.monitor.cusum
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nrt.monitor.ewma module
------------------------
-
-.. automodule:: nrt.monitor.ewma
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nrt.monitor.iqr module
-----------------------
-
-.. automodule:: nrt.monitor.iqr
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nrt.monitor.mosum module
-------------------------
-
-.. automodule:: nrt.monitor.mosum
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: nrt.monitor
-   :members:
-   :undoc-members:
-   :show-inheritance:
