+++ docs/api_reference.rst
@@ -1,0 +1,115 @@
+API Reference
+=============
+
+nrt.data package
+----------------
+
+.. automodule:: nrt.data
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.monitor package
+-------------------
+
+.. automodule:: nrt.monitor
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.monitor.ccdc module
+^^^^^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: nrt.monitor.ccdc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.monitor.cusum module
+^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: nrt.monitor.cusum
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.monitor.ewma module
+^^^^^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: nrt.monitor.ewma
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.monitor.iqr module
+^^^^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: nrt.monitor.iqr
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.monitor.mosum module
+^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: nrt.monitor.mosum
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.fit\_methods module
+-----------------------
+
+.. automodule:: nrt.fit_methods
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.log module
+--------------
+
+.. automodule:: nrt.log
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.outliers module
+-------------------
+
+.. automodule:: nrt.outliers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.stats module
+----------------
+
+.. automodule:: nrt.stats
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.utils module
+----------------
+
+.. automodule:: nrt.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+nrt.utils\_efp module
+---------------------
+
+.. automodule:: nrt.utils_efp
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: nrt
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
