### 2024-07-29 30164c9
Various minor changes to clean up doc build log of all warnings Fixes #29
