+++ docs/conf.py
@@ -58,7 +58,7 @@ templates_path = ['_templates']
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = []
+exclude_patterns = ['gallery/README.rst']
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -71,8 +71,8 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['images']
-html_logo = "images/logo.png"
+html_static_path = ['_static']
+html_logo = "_static/logo.png"
 html_theme_options = {
     'logo_only': True,
     'display_version': False,
