+++ docs/api_reference.rst
@@ -112,4 +112,5 @@ Module contents
    :members:
    :undoc-members:
    :show-inheritance:
+   :noindex:
 
