### 2024-07-29 50539da
Update gh action versions used to latest Fixes #26
