+++ .github/workflows/build_and_test.yml
@@ -13,13 +13,13 @@ jobs:
 
     steps:
       - name: checkout repo + LFS files
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
         with:
           lfs: true
       - name: checkout lfs objects
         run: git lfs checkout
       - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies and package
@@ -38,12 +38,12 @@ jobs:
     if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags/')
     steps:
       - name: Checkout repository + LFS files
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
         with:
           lfs: true
       
       - name: Set up Python 3.11
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v5
         with:
           python-version: '3.11'
 
@@ -56,7 +56,7 @@ jobs:
           python -m build
 
       - name: Publish package to PyPI
-        uses: pypa/gh-action-pypi-publish@v1.4.2
+        uses: pypa/gh-action-pypi-publish@v1.9.0
         with:
           password: ${{ secrets.PYPI_API_TOKEN }}
           user: __token__
