### 2024-07-29 edaadba
Add intro paragraph on landing page fixes #31
