+++ docs/index.rst
@@ -6,6 +6,8 @@
 Welcome to nrt's documentation!
 ===============================
 
+`nrt` is a Python package designed for near real-time detection of changes in spatio-temporal datasets, with a particular focus on monitoring forest disturbances from satellite image time-series. It offers a standardized API inspired by scikit-learn, ensuring seamless interoperability and comparison across various state-of-the-art monitoring algorithms. Optimized for rapid computation, `nrt` is suitable for operational deployment at scale. This package is an essential tool for researchers and practitioners aiming for timely and efficient monitoring, contributing to climate change mitigation, biodiversity conservation, and natural heritage preservation. Explore the official documentation to get started with installation, usage examples, and more.
+
 .. automodule:: nrt
 
 .. toctree::
