### 2024-07-29 6f3d749
pin exact versions of docs dependencies for reproducible rtd builds fixes #23
