### 2024-07-30 #6815 @AdamRJensen
@loicdtx could you make a new release and upload to Zenodo. I'll need the latest package version and the Zenodo DOI. Thanks.
