### 2024-07-30 #6815 @loicdtx
@AdamRJensen I pushed this release last week https://github.com/ec-jrc/nrt/releases/tag/v0.2.1

The zenodo project archive is pending approval of the EU funded project community https://zenodo.org/uploads/12799278
