### 2024-07-30 #6815 @AdamRJensen
> The zenodo project archive is pending approval of the EU funded project community https://zenodo.org/uploads/12799278

@loicdtx OK let me know once it is approved and we can proceed with the submission
