### 2024-08-13 #6815 @loicdtx
@AdamRJensen the repo finally got approved on Zenodo (https://zenodo.org/records/12799278). There were a couple of changes (mostly to documentation) since that release. Nothing that would justify a full version release, but I could make a [post release](https://peps.python.org/pep-0440/#post-releases) to pypi and Zenodo if you think that's necessary.
