### 2024-08-15 #6815 @kthyng
Congratulations on your new publication @loicdtx! Many thanks to @AdamRJensen and to reviewers @alexgleith and @szwiep for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.

@loicdtx Please sign up to review at https://reviewers.joss.theoj.org/ if you're interested!
